rename copySupplementalPhETIOFiles, https://github.com/phetsims/phet-io/issues/1361
